API/update to v2.6.0

+++ src/main/java/org/trebol/services/CustomerPersonRelationService.java
@@ -11,5 +11,5 @@ import org.trebol.api.pojo.CustomerPojo;
  */
 public interface CustomerPersonRelationService {
   @Nullable
-  CustomerPojo getCustomerFromPersonId(int personId);
+  CustomerPojo getCustomerFromPersonId(long personId);
 }
